Some files has been returned back.

+++ plugins/eu.modelwriter.marker.command/src/eu/modelwriter/marker/command/DeleteHandler.java
@@ -41,7 +41,6 @@ import org.eclipse.ui.texteditor.ITextEditor;
 
 import eu.modelwriter.configuration.internal.AlloyUtilities;
 import eu.modelwriter.marker.internal.AnnotationFactory;
-import eu.modelwriter.marker.internal.MappingUtilities;
 import eu.modelwriter.marker.internal.MarkUtilities;
 import eu.modelwriter.marker.internal.MarkerFactory;
 import eu.modelwriter.marker.ui.internal.views.visualizationview.Visualization;
@@ -103,15 +102,12 @@ public class DeleteHandler extends AbstractHandler {
               MarkerFactory.findMarkersByGroupId(this.file, markerGroupId);
 
           for (int i = markers.size() - 1; i >= 0; i--) {
-            if (AlloyUtilities.isExists())
-              this.deleteFromAlloyXML(markers.get(i));
+            this.deleteFromAlloyXML(markers.get(i));
             AnnotationFactory.removeAnnotation(markers.get(i));
             markers.get(i).delete();
           }
         } else {
-          MappingUtilities.removeLocation(marker);
-          if (AlloyUtilities.isExists())
-            this.deleteFromAlloyXML(this.marker);
+          this.deleteFromAlloyXML(this.marker);
           AnnotationFactory.removeAnnotation(this.marker);
           this.marker.delete();
         }
@@ -128,23 +124,23 @@ public class DeleteHandler extends AbstractHandler {
 
   @Override
   public Object execute(final ExecutionEvent event) throws ExecutionException {
-    // if (AlloyUtilities.isExists()) {
-    this.file = PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage()
-        .getActiveEditor().getEditorInput().getAdapter(IFile.class);
-    this.selection =
-        PlatformUI.getWorkbench().getActiveWorkbenchWindow().getSelectionService().getSelection();
-    this.candidateToTypeChanging = new ArrayList<IMarker>();
-    this.marker = this.getMarkerFromEditor();
-    this.editor = MarkerFactory.getOpenEditorOfMarker(this.marker);
+    if (AlloyUtilities.isExists()) {
+      this.file = PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage()
+          .getActiveEditor().getEditorInput().getAdapter(IFile.class);
+      this.selection =
+          PlatformUI.getWorkbench().getActiveWorkbenchWindow().getSelectionService().getSelection();
+      this.candidateToTypeChanging = new ArrayList<IMarker>();
+      this.marker = this.getMarkerFromEditor();
+      this.editor = MarkerFactory.getOpenEditorOfMarker(this.marker);
 
-    this.deleteMarker();
-    this.refresh();
-    // } else {
-    // final MessageDialog infoDialog = new MessageDialog(new Shell(), "System Information", null,
-    // "You dont have any registered alloy file to system.", MessageDialog.INFORMATION,
-    // new String[] {"OK"}, 0);
-    // infoDialog.open();
-    // }
+      this.deleteMarker();
+      this.refresh();
+    } else {
+      final MessageDialog infoDialog = new MessageDialog(new Shell(), "System Information", null,
+          "You dont have any registered alloy file to system.", MessageDialog.INFORMATION,
+          new String[] {"OK"}, 0);
+      infoDialog.open();
+    }
     return null;
   }
 
@@ -152,22 +148,16 @@ public class DeleteHandler extends AbstractHandler {
    * @param selectedMarker from text
    */
   private void findCandidateToTypeChangingMarkers(final IMarker selectedMarker) {
-    // final Map<IMarker, String> fieldsSources =
-    // AlloyUtilities.getRelationsOfSecondSideMarker(selectedMarker);
-    // final ArrayList<IMarker> relationsSources =
-    // AlloyUtilities.getSourcesOfMarkerAtRelations(selectedMarker);
-    //
-    // for (final IMarker iMarker : fieldsSources.keySet()) {
-    // this.candidateToTypeChanging.add(iMarker);
-    // }
-    //
-    // for (final IMarker iMarker : relationsSources) {
-    // this.candidateToTypeChanging.add(iMarker);
-    // }
+    final Map<IMarker, String> fieldsSources =
+        AlloyUtilities.getRelationsOfSecondSideMarker(selectedMarker);
+    final ArrayList<IMarker> relationsSources =
+        AlloyUtilities.getSourcesOfMarkerAtRelations(selectedMarker);
 
-    final ArrayList<IMarker> sources = MappingUtilities.getSourcesOfMarker(selectedMarker);
+    for (final IMarker iMarker : fieldsSources.keySet()) {
+      this.candidateToTypeChanging.add(iMarker);
+    }
 
-    for (IMarker iMarker : sources) {
+    for (final IMarker iMarker : relationsSources) {
       this.candidateToTypeChanging.add(iMarker);
     }
   }
